requirements.txt
tensorflow-cpu>=2.6.0
numpy>=1.19.2
Pillow>=8.0.1
streamlit>=1.0.0
